Reformatted testing folder, combined spec and tester files into tester
